pyproject.toml
[project]
name = "polymarket-autoresearch"
version = "0.1.0"
description = "Autonomous experiment loop for Polymarket paper trading strategies"
requires-python = ">=3.10"
dependencies = [
    "requests>=2.31.0",
    "rich>=13.7.0",
    "httpx>=0.27.0",
    "websockets>=12.0",
    "click>=8.1.7",
    "tabulate>=0.9.0",
    "numpy>=1.26.0",
    "pandas>=2.1.0",
    "pyyaml>=6.0.1",
    "numpy",
    "pandas",
    "requests",
]

[project.optional-dependencies]
dev = ["pytest>=7.4.0", "pytest-asyncio>=0.23.0", "black>=23.0.0", "ruff>=0.1.0"]
